requirements.txt
discord.py>=2.3.2
google-generativeai>=0.4.0
python-dotenv>=1.0.0
pymongo>=4.6.0
aiohttp>=3.9.0